move/rename 'parallel' module to scoop_framework package, name it 'util'
