+++ rmgpy/scoop_framework/util.py
